pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.135.1",
    "google-adk>=1.27.0",
    "httpx>=0.28.1",
    "uvicorn>=0.41.0",
]
